DEVOPS/pyproject.toml
[project]
name = "ai-devops-team"
version = "0.1.0"
description = "AI DevOps Agent Team using CrewAI for professional, automated infrastructure management."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.28.8",
    "crewai-tools>=0.1.6",
    "langchain-openai>=0.1.3",
    "langchain-anthropic>=0.1.11",
    "python-dotenv>=1.0.1",
    "pydantic>=2.7.0"
]

[project.scripts]
devops = "main:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
